+++ weblate/appsettings.py
@@ -132,6 +132,7 @@ CHECK_LIST = getvalue('CHECK_LIST', (
     'weblate.trans.checks.format.JavascriptFormatCheck',
     'weblate.trans.checks.angularjs.AngularJSInterpolationCheck',
     'weblate.trans.checks.consistency.PluralsCheck',
+    'weblate.trans.checks.consistency.SamePluralsCheck',
     'weblate.trans.checks.consistency.ConsistencyCheck',
     'weblate.trans.checks.chars.NewlineCountingCheck',
     'weblate.trans.checks.markup.BBCodeCheck',