+++ docs/changes.rst
@@ -6,6 +6,8 @@ weblate 2.10
 
 Released on ? 2016.
 
+* Added quality check to check whether plurals are translated differently.
+
 weblate 2.9
 -----------